n2o-aeration-model/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "n2o-aeration-model"
version = "0.2.0"
description = "Semi-mechanistic ODE model of aeration strategy and N2O dynamics in secondary wastewater treatment"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "OpenAI"}]
dependencies = [
  "numpy>=1.26",
  "scipy>=1.12",
  "pandas>=2.2",
  "matplotlib>=3.8",
  "plotly>=5.22",
  "streamlit>=1.36",
  "streamlit-autorefresh>=1.0.1",
  "PyYAML>=6.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
